Revert "[#14] 상품 등록 기능 구현"
