+++ src/main/java/com/project/smartstore/service/ProductService.java
@@ -4,22 +4,11 @@ import com.project.smartstore.dto.ProductDto;
 import com.project.smartstore.dto.ProductListDto;
 import com.project.smartstore.dto.SearchConditionDto;
 import java.util.List;
-import com.project.smartstore.mapper.ProductMapper;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
 
 
-@Service
-@RequiredArgsConstructor
-public class ProductService {
+public interface ProductService {
 
-  private final ProductMapper productMapper;
+  void addProductInStore(ProductDto productDto);
 
-  public void addProductInStore(ProductDto productDto) {
-    productMapper.insertProduct(productDto);
-  }
-
-  public List<ProductListDto> getProductList(SearchConditionDto searchConditionDto) {
-    return productMapper.selectProductList(searchConditionDto);
-  }
+  List<ProductListDto> getProductList(SearchConditionDto searchConditionDto);
 }
